Repeat and Missing Numbers O(N) - Type Casting neccessary

+++ Arrays/Repeat And Missign Numbers.cpp
@@ -1,19 +1,19 @@
 vector<int> Solution::repeatedNumber(const vector<int> &A) 
 {
-    int n = A.size();
+    long n = A.size();
 
-    long long s = (n*(n+1))/2;
-    long long ss = (n*(n+1)*(2*n+1))/6;
+    long s = (n*(n+1))/2;
+    long ss = (n*(n+1)*(2*n+1))/6;
     
-    long long a = 0 ;
+    int a = 0 ;
     
     for(int i = 0 ; i<n ; i++)
     {
-        s -= A[i];
-        ss -= A[i]*A[i];
+        s -= (long)A[i];
+        ss -= ((long)A[i]*(long)A[i]);
     }
 
-    a = ((ss/s) - s)/2 ;
+    a = (int)(((ss/s) - s)/2) ;
     
-    return {a, a+s};
+    return {a, a+(int)s};
 }
